0.8.20.29: 	Merge SB-FUTEX and SB-THREAD: the latter now requires the 	former.  SBCL threads now require Linux kernel 2.6, or an NPTL 	backport to 2.4 such as the Red Hat one
	Lock/unlock functions take long (not int) as arguments: this makes
	a difference on 64 bit ports (or would do if we had threading support
	on either of them)

+++ src/runtime/mips-arch.h
@@ -3,7 +3,7 @@
 
 
 static inline void 
-get_spinlock(lispobj *word,int value)
+get_spinlock(lispobj *word,long value)
 {
     *word=value;		/* FIXME for threads */
 }
