+++ src/runtime/sparc-arch.h
@@ -2,7 +2,7 @@
 #define _SPARC_ARCH_H
 
 static inline void 
-get_spinlock(lispobj *word,int value)
+get_spinlock(lispobj *word,long value)
 {
     *word=value;		/* FIXME for threads */
 }
